refactor(agent): split session/builder.rs (1789 lines) into builder/ directory
Sub-modules: factory (from_config constructors), setters (fluent API),
tools (tool registration), inner (build finalization + delegation setup).
